[EventDispatcher][VarDumper] optimize perf by leveraging Closure::fromCallable()
