+++ src/Symfony/Component/EventDispatcher/EventDispatcher.php
@@ -30,6 +30,14 @@ class EventDispatcher implements EventDispatcherInterface
 {
     private $listeners = array();
     private $sorted = array();
+    private $optimized;
+
+    public function __construct()
+    {
+        if (__CLASS__ === \get_class($this)) {
+            $this->optimized = array();
+        }
+    }
 
     /**
      * {@inheritdoc}
@@ -40,7 +48,13 @@ class EventDispatcher implements EventDispatcherInterface
             $event = new Event();
         }
 
-        if ($listeners = $this->getListeners($eventName)) {
+        if (null !== $this->optimized && null !== $eventName) {
+            $listeners = $this->optimized[$eventName] ?? (empty($this->listeners[$eventName]) ? array() : $this->optimizeListeners($eventName));
+        } else {
+            $listeners = $this->getListeners($eventName);
+        }
+
+        if ($listeners) {
             $this->doDispatch($listeners, $eventName, $event);
         }
 
@@ -86,11 +100,10 @@ class EventDispatcher implements EventDispatcherInterface
             $listener[0] = $listener[0]();
         }
 
-        foreach ($this->listeners[$eventName] as $priority => $listeners) {
-            foreach ($listeners as $k => $v) {
+        foreach ($this->listeners[$eventName] as $priority => &$listeners) {
+            foreach ($listeners as &$v) {
                 if ($v !== $listener && \is_array($v) && isset($v[0]) && $v[0] instanceof \Closure) {
                     $v[0] = $v[0]();
-                    $this->listeners[$eventName][$priority][$k] = $v;
                 }
                 if ($v === $listener) {
                     return $priority;
@@ -123,7 +136,7 @@ class EventDispatcher implements EventDispatcherInterface
     public function addListener($eventName, $listener, $priority = 0)
     {
         $this->listeners[$eventName][$priority][] = $listener;
-        unset($this->sorted[$eventName]);
+        unset($this->sorted[$eventName], $this->optimized[$eventName]);
     }
 
     /**
@@ -139,21 +152,17 @@ class EventDispatcher implements EventDispatcherInterface
             $listener[0] = $listener[0]();
         }
 
-        foreach ($this->listeners[$eventName] as $priority => $listeners) {
-            foreach ($listeners as $k => $v) {
+        foreach ($this->listeners[$eventName] as $priority => &$listeners) {
+            foreach ($listeners as $k => &$v) {
                 if ($v !== $listener && \is_array($v) && isset($v[0]) && $v[0] instanceof \Closure) {
                     $v[0] = $v[0]();
                 }
                 if ($v === $listener) {
-                    unset($listeners[$k], $this->sorted[$eventName]);
-                } else {
-                    $listeners[$k] = $v;
+                    unset($listeners[$k], $this->sorted[$eventName], $this->optimized[$eventName]);
                 }
             }
 
-            if ($listeners) {
-                $this->listeners[$eventName][$priority] = $listeners;
-            } else {
+            if (!$listeners) {
                 unset($this->listeners[$eventName][$priority]);
             }
         }
@@ -215,22 +224,46 @@ class EventDispatcher implements EventDispatcherInterface
 
     /**
      * Sorts the internal list of listeners for the given event by priority.
-     *
-     * @param string $eventName The name of the event
      */
-    private function sortListeners($eventName)
+    private function sortListeners(string $eventName)
     {
         krsort($this->listeners[$eventName]);
         $this->sorted[$eventName] = array();
 
-        foreach ($this->listeners[$eventName] as $priority => $listeners) {
+        foreach ($this->listeners[$eventName] as &$listeners) {
             foreach ($listeners as $k => $listener) {
                 if (\is_array($listener) && isset($listener[0]) && $listener[0] instanceof \Closure) {
                     $listener[0] = $listener[0]();
-                    $this->listeners[$eventName][$priority][$k] = $listener;
                 }
                 $this->sorted[$eventName][] = $listener;
             }
         }
+    }
+
+    /**
+     * Optimizes the internal list of listeners for the given event by priority.
+     */
+    private function optimizeListeners(string $eventName): array
+    {
+        krsort($this->listeners[$eventName]);
+        $this->optimized[$eventName] = array();
+
+        foreach ($this->listeners[$eventName] as &$listeners) {
+            foreach ($listeners as &$listener) {
+                $closure = &$this->optimized[$eventName][];
+                if (\is_array($listener) && isset($listener[0]) && $listener[0] instanceof \Closure) {
+                    $closure = static function (...$args) use (&$listener, &$closure) {
+                        if ($listener[0] instanceof \Closure) {
+                            $listener[0] = $listener[0]();
+                        }
+                        ($closure = \Closure::fromCallable($listener))(...$args);
+                    };
+                } else {
+                    $closure = $listener instanceof \Closure ? $listener : \Closure::fromCallable($listener);
+                }
+            }
+        }
+
+        return $this->optimized[$eventName];
     }
 }
