+++ src/Symfony/Component/VarDumper/Cloner/AbstractCloner.php
@@ -176,7 +176,10 @@ abstract class AbstractCloner implements ClonerInterface
     public function addCasters(array $casters)
     {
         foreach ($casters as $type => $callback) {
-            $this->casters[strtolower($type)][] = \is_string($callback) && false !== strpos($callback, '::') ? explode('::', $callback, 2) : $callback;
+            $closure = &$this->casters[strtolower($type)][];
+            $closure = $callback instanceof \Closure ? $callback : static function (...$args) use ($callback, &$closure) {
+                return ($closure = \Closure::fromCallable($callback))(...$args);
+            };
         }
     }
